requirements.txt
Flask==3.1.0
functions_framework==3.8.2
protobuf==5.29.2
python-dotenv==1.0.1
Requests==2.32.3
Werkzeug==3.1.3
